added FFTW3 library, transferred time to ms, developed ideas to talk about for future

+++ include/FFT.h
@@ -1,4 +1,5 @@
 #pragma once //prevent file from being included more than once
 #include <Arduino.h>
 
-void fft(double xval, double yval, double zval, int time);
+/* *************** DECLARE FUNCTIONS HERE *************** */
+void fft(double xval, double yval, double zval, long time);
